pyproject.toml
[tool.poetry]
name = "arbbot"
version = "0.1.0"
description = ""
authors = ["zenix"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
ccxt = "^4.1.51"
python-dotenv = "^1.0.0"
ruamel-yaml = "^0.18.5"
aiohttp = "^3.9.0"
pyglet = "^2.0.10"
colorama = "^0.4.6"
coloredlogs = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
